requirements.txt
apns2==0.7.2
asgiref==3.4.1
autopep8==2.0.1
Babel==2.11.0
boto3==1.16.22
botocore==1.19.56
CacheControl==0.12.11
cachetools==5.2.1
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==2.0.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.4
deprecation==2.1.0
Django==3.2.8
django-filter==21.1
pywebpush==1.9.4
django-push-notifications==3.0.0
django-storages==1.12
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
drf-yasg==1.20.0
firebase-admin==5.1.0
google-api-core==2.11.0
google-api-python-client==2.73.0
google-auth==2.16.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.9.0
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
grpcio==1.51.1
grpcio-status==1.51.1
h2==2.6.2
hpack==3.0.0
http-ece==1.1.0
httplib2==0.21.0
hyper==0.7.0
hyperframe==3.2.0
idna==3.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.2
jmespath==0.10.0
MarkupSafe==2.0.1
msgpack==1.0.4
packaging==21.0
Pillow==9.4.0
proto-plus==1.22.2
protobuf==4.21.12
psycopg2-binary==2.9.5
py-vapid==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybase64==1.2.3
pycodestyle==2.10.0
pycparser==2.21
PyJWT==1.7.1
pyotp==2.3.0
pyparsing==2.4.7
PyPDF2==2.11.1
python-dateutil==2.8.2
python-decouple==3.7
python-http-client==3.3.7
pytz==2021.3
requests==2.26.0
rsa==4.9
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
s3transfer==0.3.7
sendgrid==6.9.2
six==1.15.0
sqlparse==0.4.2
starkbank-ecdsa==2.2.0
tomli==2.0.1
uritemplate==4.0.0
urllib3==1.26.7
uuid==1.30
python-dotenv==0.21.1
django-crontab==0.7.1
django-cors-headers==3.13.0
twilio==6.63.
websocket-client==0.57.0
websockets==8.1